requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
plotly>=5.18.0
fpdf2>=2.7.0
kaleido==0.2.1
